requirements.txt
fastapi
uvicorn
requests
leia-br
pydantic
